requirements.txt
Flask==2.0.1
gspread==3.6.0
google-auth==1.32.1
requests==2.25.1
